+++ frontend/atualizar.css
@@ -1,7 +1,7 @@
 /*Google Fontes*/
 @import url('https://fonts.googleapis.com/css2?family=Lora:ital,wght@0,400..700;1,400..700&family=Poppins:ital,wght@0,100;0,200;0,300;0,400;0,500;0,600;0,700;0,800;0,900;1,100;1,200;1,300;1,400;1,500;1,600;1,700;1,800;1,900&display=swap');
 *{
-  font-family: "Lora", serif;
+    font-family: "Lora", "Poppins";
     margin: 0;
     padding: 0;
     box-sizing: border-box;
@@ -94,7 +94,7 @@ header{
 }
 
 .iconesNav {
-  position: fixed;           
+  display: fixed;           
   justify-content: center; 
   align-items: center;     
   overflow: visible;
