+++ frontend/atualizar.html
@@ -1,0 +1,116 @@
+<!DOCTYPE html>
+<html lang="pt-br">
+<head>
+
+    <meta charset="UTF-8">
+    
+    <title>Saborvete - Cadastro</title>
+    
+    <meta name="viewport" content="width=device-width, initial-
+    scalle=1.0">
+    
+    <link rel="stylesheet" href="atualizar.css">
+  
+      <!--Link para site de ícones-->
+      <link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/boxicons@latest/css/boxicons.min.css">
+    
+</head>
+<body>
+          <!--Barra de navegação-->
+          <header>
+            <a href="home.html" class="logo">
+              <img src="/frontend/Imagens/logoSaborvete-semFundo.png" alt="Saborvete">
+            </a>
+          
+            <!-- Ícones atualizados com direcionamento -->
+            <div class="navbar">
+              <a href="/frontend/carrinho.html" class="iconesNav"><i class="bx bx-cart" id="carrinho-icon"></i></a>
+              <a href="/frontend/atualizar.html" class="iconesNav"><i class="bx bx-menu" id="menu-icon"></i></a>
+            </div>
+          </header>      
+      
+    <h1>CADASTRO</h1>
+
+    <nav class="navbar"></nav>
+
+    <form action="atualizar.html" method="get" class="form">
+        <h2>CADASTRO</h2>
+    
+    <!-- Dropdown de sorvetes -->
+     <div class="campo">
+        <label for="lista-sorvetes">Produto:</label>
+        <select id="lista-sorvetes">
+            <option value="">Escolha um produto...</option>
+        </select>
+    </div>
+
+    <div class="campo">
+        <label for="preco">Preço:</label>
+        <input type="number" id="preco">
+    </div>
+    
+    <div class="campo">
+        <label for="quantidade">Quantidade:</label>
+        <input type="number" id="quantidade">
+    </div>
+
+    <button type="submit" class="btn-salvar" id="btn-salvar">Salvar</button>
+
+    <script>
+    document.addEventListener("DOMContentLoaded", async () => {
+        const selectSorvetes = document.getElementById("lista-sorvetes");
+
+        try {
+            const resposta = await fetch("http://localhost:3001/sorvetes");
+            if (!resposta.ok) throw new Error("Erro na requisição");
+
+            const sorvetes = await resposta.json();
+
+            // Preenchendo o dropdown
+            sorvetes.forEach(sorvete => {
+                const option = document.createElement("option");
+                option.value = sorvete.nome;
+                option.textContent = sorvete.nome;
+                selectSorvetes.appendChild(option);
+            });
+        } catch (erro) {
+            console.error("Erro ao buscar sorvetes:", erro);
+        }
+    });
+
+    document.getElementById("btn-salvar").addEventListener("click", async () => {
+        const nome = document.getElementById("lista-sorvetes").value;
+        const preco = document.getElementById("preco").value;
+        const quantidade = document.getElementById("quantidade").value;
+
+        if (!nome || preco === "" || quantidade === "") {
+            alert("Preencha todos os campos!");
+            return;
+        }
+
+        try {
+            const resposta = await fetch("http://localhost:3001/atualizar", {
+                method: "PUT",
+                headers: { "Content-Type": "application/json" },
+                body: JSON.stringify({ nome, preco: Number(preco), quantidade: Number(quantidade) })
+            });
+
+            const resultado = await resposta.json();
+            alert(resultado.mensagem || resultado.erro);
+        } catch (erro) {
+            console.error("Erro ao atualizar:", erro);
+            alert("Erro ao atualizar o produto.");
+        }
+    });
+    </script>
+
+    <footer class="footer">
+        <p class="copyright">&copy; 2025 - Todos os direitos reservados</p>
+    </footer>
+  
+    <!--Linkando o JavaScript-->
+      <script src="cadastro.js"></script>
+      <script src="https://code.jquery.com/jquery-3.6.0.min.js"></script>
+  
+</body>
+</html>
