--- frontend/cadastro.html
@@ -1,35 +1,0 @@
-<!DOCTYPE html>
-<html lang="pt-br">
-<head>
-    <meta charset="UTF-8">
-    <title>Saborvete - Cadastro</title>
-    <link rel="icon" href="Imagens/abaLogo.png" type="image/png">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <link rel="stylesheet" href="atualizar.css">
-    <link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/boxicons@latest/css/boxicons.min.css">
-</head>
-<body>
-    <header>
-        <a href="home.html" class="logo">
-            <img src="Imagens/logoSaborvete-semFundo.png" alt="Saborvete">
-        </a>
-        <div class="navbar">
-            <a href="carrinho.html" class="iconesNav"><i class="bx bx-cart"></i></a>
-            <a href="menu.html" class="iconesNav"><i class="bx bx-menu"></i></a>
-        </div>
-    </header>
-
-    <h1>CADASTRO</h1>
-    <form class="form">
-        <h2>Cadastro de Produto</h2>
-
-        <div class="campo">
-            <label for="produto">Produto:</label>
-            <select id="produto">
-                <option value="">Escolha um produto...</option>
-            </select>
-        </div>
-
-        <div class="campo">
-            <label for="preco">Preço:</label>
-            <input type="number" id="preco" step="0.01" min="0">
