+++ frontend/carrinho.html
@@ -9,7 +9,7 @@
 
   <title>Saborvete - Carrinho</title>
 
-  <link rel="icon" href="images/abaLogo.png" type="image/png">
+  <link rel="icon" href="/frontend/Imagens/abaLogo.png" type="image/png">
 
   <link rel="stylesheet" href="carrinho.css">
 
@@ -21,13 +21,13 @@
     <!--Barra de navegação-->
     <header>
       <a href="home.html" class="logo">
-        <img src="images/logoSaborvete-semFundo.png" alt="Saborvete">
+        <img src="/frontend/Imagens/logoSaborvete-semFundo.png" alt="Saborvete">
       </a>
 
       <!--pngs foram trocados por icones-->
       <div class="navbar">
-        <a href="#" class="iconesNav" target="_blank"><i class="bx bx-cart" id="carrinho-icon"></i></a>
-        <a href="#" class="iconesNav" target="_blank"><i class="bx bx-menu" id="menu-icon"></i></a>
+        <a href="/frontend/carrinho.html" class="iconesNav" target="_blank"><i class="bx bx-cart" id="carrinho-icon"></i></a>
+        <a href="/frontend/atualizar.html" class="iconesNav" target="_blank"><i class="bx bx-menu" id="menu-icon"></i></a>
       </div>
     </header>
