+++ frontend/home.html
@@ -14,7 +14,7 @@
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
 
     <!--Icone da aba-->
-    <link rel="icon" href="Imagens/abaLogo.png" type="image/png">
+    <link rel="icon" href="/frontend/Imagens/abaLogo.png" type="image/png">
     
     <title>Saborvete</title>
 
@@ -30,13 +30,13 @@
         <!--Barra de navegação-->
         <header>
             <a href="home.html" class="logo">
-                <img src="Imagens/logoSaborvete-semFundo.png" alt="Saborvete">
+                <img src="/frontend/Imagens/logoSaborvete-semFundo.png" alt="Saborvete">
             </a>
 
             <!--Icones de menu-->
             <div class="navbar">
-                <a href="/carrinho/index.html" class="iconesNav" target="_blank"><i class="bx bx-cart" id="carrinho-icon"></i></a>
-                <a href="#" class="iconesNav" target="_blank"><i class="bx bx-menu" id="menu-icon"></i></a>
+                <a href="/frontend/carrinho.html" class="iconesNav" target="_blank"><i class="bx bx-cart" id="carrinho-icon"></i></a>
+                <a href="/frontend/atualizar.html" class="iconesNav" target="_blank"><i class="bx bx-menu" id="menu-icon"></i></a>
             </div>
         </header>
